refactor: auto-split test_energy_monitor.py
